split full function constellation framework contract into operations, queries, events and tests rs
