+++ contracts/option/src/lib.rs
@@ -1,4 +1,9 @@
 mod contracts;
 pub mod msg;
 pub mod state;
+pub mod events;
+pub mod query;
+pub mod operations;
+#[cfg(test)]
+pub mod tests;
 mod error;
